requirements.txt
beautifulsoup4
lxml
pandas
python-dateutil
requests
tabulate
